services/iris/endpoint/requirements.txt
flask
numpy
scipy
scikit-learn
grpcio
grpcio-tools
googleapis-common-protos